refactor: import v1 setupClickToMoveCursor instead of extracting it
Agent-Logs-Url: https://github.com/yashasg/superset/sessions/b8820492-651a-426c-8ec9-57ee22c4b712
